refactor: split main.rs into cmd/ submodules (Issue 2)
main.rs drops from 799 lines to 94 — just main(), the run()
dispatcher, and module declarations. All command implementations
move to a new cmd/ submodule:

  cmd/cmd_loop.rs  — run_loop, LoopObserver, and all loop helpers
                     (bind_socket_listener, install_shutdown_handler,
                     finalize_shutdown, resolve_tick_dt,
                     sleep_until_next_deadline, sleep_with_socket,
                     read_pv_with_escalation, log_interval_slip_if_needed)
  cmd/cmd_once.rs  — run_once, resolve_once_dt, clamp_once_wall_clock_dt
  cmd/cmd_pipe.rs  — run_pipe
  cmd/cmd_state.rs — run_status_dispatch, run_status, run_purge, run_init
  cmd/cmd_socket.rs — socket_send_and_print, run_socket_{hold,resume,reset,save,set}
  cmd/mod.rs       — module declarations, open_log, pub(crate) re-exports

Also fixes a pre-existing test reference crate::app::IterationRecord
→ pid_ctl::app::IterationRecord in tune/tests.rs (was only resolving
because the old main.rs pulled pid_ctl::app into the crate root).
